complete refactoring of createActorVertices in ActorClass instances
