+++ 2025-07-12-pong/app/Main.hs
@@ -99,76 +99,6 @@ isQuitEvent e = case e of
     keysymKeycode (keyboardEventKeysym keyboardEvent) == KeycodeEscape
   _ -> False
 
-createActorVertices :: Actor -> IO (GL.VertexArrayObject, GL.BufferObject, Int)
-createActorVertices (Ball ballPos size _) = do
-  let V2 ballX ballY = ballPos
-      ballRadius = size / 2
-      vertices
-        = V.fromList
-        $ concat [ [ballX, ballY, 0.0]
-                 ] ++ unwrapV2 (createCircleVertices ballPos ballRadius)
-      numVertices = (V.length vertices) `div` 3
-
-  vao <- GL.genObjectName
-  GL.bindVertexArrayObject $= Just vao
-
-  vbo <- GL.genObjectName
-  GL.bindBuffer GL.ArrayBuffer $= Just vbo
-  V.unsafeWith vertices $ \ptr ->
-    GL.bufferData GL.ArrayBuffer $= (fromIntegral (V.length vertices * 4), ptr, GL.StaticDraw)
-
-  GL.vertexAttribArray (GL.AttribLocation 0) $= GL.Enabled
-  GL.vertexAttribPointer (GL.AttribLocation 0) $=
-    (GL.ToFloat, GL.VertexArrayDescriptor 3 GL.Float 0 nullPtr)
-
-  GL.bindBuffer GL.ArrayBuffer $= Nothing
-  GL.bindVertexArrayObject $= Nothing
-
-  return (vao, vbo, numVertices)
-
-  where
-    unwrapV2 :: [V2 Float] -> [Float]
-    unwrapV2 [] = []
-    unwrapV2 ((V2 x y):xs) = x : y : 0.0 : unwrapV2 xs
-createActorVertices (Paddle pos width height) = do
-  let V2 rectX rectY = pos
-      V2 rectW rectH = V2 width height
-
-      vertices = V.fromList $
-        concatMap (\(V2 x y) -> [x, y, 0.0]) $
-        [ V2 rectX         rectY
-        , V2 (rectX+rectW) rectY
-        , V2 rectX         (rectY+rectH)
-        , V2 rectX         (rectY+rectH)
-        , V2 (rectX+rectW) rectY
-        , V2 (rectX+rectW) (rectY+rectH)
-        ]
-
-      numVertices = V.length vertices `div` 3
-
-  vao <- GL.genObjectName
-  GL.bindVertexArrayObject $= Just vao
-
-  vbo <- GL.genObjectName
-  GL.bindBuffer GL.ArrayBuffer $= Just vbo
-  V.unsafeWith vertices $ \ptr ->
-    GL.bufferData GL.ArrayBuffer $= (fromIntegral (V.length vertices * 4), ptr, GL.StaticDraw)
-
-  GL.vertexAttribArray (GL.AttribLocation 0) $= GL.Enabled
-  GL.vertexAttribPointer (GL.AttribLocation 0) $=
-    (GL.ToFloat, GL.VertexArrayDescriptor 3 GL.Float 0 nullPtr)
-
-  GL.bindBuffer GL.ArrayBuffer $= Nothing
-  GL.bindVertexArrayObject $= Nothing
-
-  return (vao, vbo, numVertices)
-
-createCircleVertices :: V2 Float -> Float -> [V2 Float]
-createCircleVertices pos radius = (calculateCircumferencePoint pos radius) <$> [1..360]
-
-calculateCircumferencePoint :: V2 Float -> Float -> Int -> V2 Float
-calculateCircumferencePoint pos r th = V2 (r * (sin (fromIntegral th))) (r * (cos (fromIntegral th))) + pos
-
 createTriangleVertices :: IO (GL.VertexArrayObject, GL.BufferObject)
 createTriangleVertices = do
   let vertices = V.fromList
